refactor: Spot::GoogleClient 統合（GooglePlacesAdapter + PlaceDetailsService） (#508)
- GooglePlacesAdapter と PlaceDetailsService を Spot::GoogleClient に統合
- HTTP処理・APIキー・エラーハンドリングを共通化
- Spot::GoogleClient.find_by_name（旧 GooglePlacesAdapter.find_place）
- Spot::GoogleClient.fetch_details（旧 PlaceDetailsService.fetch）
- Spot モデル・InfowindowsController の参照を更新
- テストを spec/models/spot/google_client_spec.rb に移動
